+++ clang/lib/Basic/IdentifierTable.cpp
@@ -178,7 +178,7 @@ static KeywordStatus getKeywordStatusHelper(const LangOptions &LangOpts,
 }
 
 KeywordStatus clang::getKeywordStatus(const LangOptions &LangOpts,
-                                      unsigned Flags) {
+                                      uint64_t Flags) {
   // KEYALL means always enabled, so special case this one.
   if (Flags == KEYALL) return KS_Enabled;
   // These are tests that need to 'always win', as they are special in that they
@@ -192,7 +192,7 @@ KeywordStatus clang::getKeywordStatus(const LangOptions &LangOpts,
   KeywordStatus CurStatus = KS_Unknown;
 
   while (Flags != 0) {
-    unsigned CurFlag = Flags & ~(Flags - 1);
+    uint64_t CurFlag = Flags & ~(Flags - 1);
     Flags = Flags & ~CurFlag;
     CurStatus = std::max(
         CurStatus,
@@ -204,7 +204,7 @@ KeywordStatus clang::getKeywordStatus(const LangOptions &LangOpts,
   return CurStatus;
 }
 
-static bool IsKeywordInCpp(unsigned Flags) {
+static bool IsKeywordInCpp(uint64_t Flags) {
   return (Flags & (KEYCXX | KEYCXX11 | KEYCXX20 | BOOLSUPPORT | WCHARSUPPORT |
                    CHAR8SUPPORT)) != 0;
 }
@@ -220,7 +220,7 @@ static void MarkIdentifierAsKeywordInCpp(IdentifierTable &Table,
 /// identifiers because they are language keywords.  This causes the lexer to
 /// automatically map matching identifiers to specialized token codes.
 static void AddKeyword(StringRef Keyword,
-                       tok::TokenKind TokenCode, unsigned Flags,
+                       tok::TokenKind TokenCode, uint64_t Flags,
                        const LangOptions &LangOpts, IdentifierTable &Table) {
   KeywordStatus AddResult = getKeywordStatus(LangOpts, Flags);
 
@@ -838,7 +838,7 @@ IdentifierTable::getFutureCompatDiagKind(const IdentifierInfo &II,
                                          const LangOptions &LangOpts) {
   assert(II.isFutureCompatKeyword() && "diagnostic should not be needed");
 
-  unsigned Flags = llvm::StringSwitch<unsigned>(II.getName())
+  uint64_t Flags = llvm::StringSwitch<uint64_t>(II.getName())
 #define KEYWORD(NAME, FLAGS) .Case(#NAME, FLAGS)
 #include "clang/Basic/TokenKinds.def"
 #undef KEYWORD
